Move message parsing to parser subpackage
