+++ ui/messages/parser/htmltagarray.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package messages
+package parser
 
 // TagWithMeta is an open HTML tag with some metadata (e.g. list index, a href value).
 type TagWithMeta struct {
